+++ src/main/java/com/yourmode/yourmodebackend/domain/survey/dto/SurveyHistoryWithAnswersResponseDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Schema(description = "설문 내역 + 답변 응답 DTO")
 public class SurveyHistoryWithAnswersResponseDto {
     @Schema(description = "설문 내역 ID", example = "1")
-    private Long historyId;
+    private Integer historyId;
     @Schema(description = "설문 응답 일시", example = "2025-07-17T12:00:00.000")
     private LocalDateTime createdAt;
     @Schema(description = "답변 목록")
