+++ src/main/java/com/yourmode/yourmodebackend/domain/user/dto/request/CommonSignupRequest.java
@@ -12,5 +12,5 @@ public interface CommonSignupRequest {
     Float getHeight();
     Float getWeight();
     Gender getGender();
-    Long getBodyTypeId();
+    Integer getBodyTypeId();
 }
