+++ src/main/java/com/yourmode/yourmodebackend/domain/user/service/UserProfileService.java
@@ -9,7 +9,7 @@ public interface UserProfileService {
      * @param userId 사용자 ID
      * @return UserProfileResponseDto
      */
-    UserProfileResponseDto getMyProfile(Long userId);
+    UserProfileResponseDto getMyProfile(Integer userId);
 
     /**
      * 현재 로그인한 사용자의 프로필 정보를 수정합니다.
@@ -17,12 +17,12 @@ public interface UserProfileService {
      * @param dto 프로필 수정 요청 DTO
      * @return UserProfileResponseDto
      */
-    UserProfileResponseDto updateMyProfile(Long userId, UserProfileUpdateRequestDto dto);
+    UserProfileResponseDto updateMyProfile(Integer userId, UserProfileUpdateRequestDto dto);
 
     /**
      * 사용자의 비밀번호를 변경합니다.
      * @param userId 사용자 ID
      * @param newPassword 새 비밀번호(암호화 전)
      */
-    void updatePassword(Long userId, String newPassword);
+    void updatePassword(Integer userId, String newPassword);
 }
